server: don't guess process name for portfile
Instead, only write the port file if the user
specifies `--portfile=ports/foo` on the CLI.

+++ server.js
@@ -6,9 +6,9 @@ import { readable as isReadableStream } from 'is-stream';
 
 const argv = minimist(process.argv.slice(2));
 
-let portfile;
 let filename = argv._.shift();
 let port = argv.port || parseInt(process.env.PORT, 10) || 0;
+let portfile = argv.portfile;
 
 if (!filename) {
   throw new Error('A server filename must be given!');
@@ -90,9 +90,9 @@ server.listen(port, (err) => {
 
   console.log('listening on port %d', port);
 
-  var name = process.env.MONGROUP_NAME;
-  portfile = 'ports/' + name;
-  write(portfile, String(port), 'ascii');
+  if (portfile) {
+    write(portfile, String(port), 'ascii');
+  }
 });
 
 // simply log unhandled rejections, like Chrome does
@@ -106,10 +106,11 @@ process.on('SIGINT', () => server.close());
 process.on('SIGQUIT', () => server.close());
 
 process.on('exit', () => {
-  try {
-    var name = process.env.MONGROUP_NAME;
-    unlink(portfile);
-  } catch (e) {
-    console.error(e.stack);
+  if (portfile) {
+    try {
+      unlink(portfile);
+    } catch (e) {
+      console.error(e.stack);
+    }
   }
 });
